Move fs/path to the top-level
