+++ fs/mod.ts
@@ -6,7 +6,6 @@ export * from "./ensure_link.ts";
 export * from "./ensure_symlink.ts";
 export * from "./exists.ts";
 export * from "./glob.ts";
-export * from "./globrex.ts";
 export * from "./move.ts";
 export * from "./copy.ts";
 export * from "./read_file_str.ts";
