+++ media_types/deps.ts
@@ -1,6 +1,6 @@
 // Copyright 2018-2019 the Deno authors. All rights reserved. MIT license.
 
-export { extname } from "../fs/path/mod.ts";
+export { extname } from "../path/mod.ts";
 
 interface DB {
   [mediaType: string]: {
