+++ mime/multipart.ts
@@ -5,11 +5,11 @@ const { min, max } = Math;
 type Closer = Deno.Closer;
 type Reader = Deno.Reader;
 type Writer = Deno.Writer;
-import { FormFile } from "../multipart/formfile.ts";
 import { equal, findIndex, findLastIndex, hasPrefix } from "../bytes/mod.ts";
-import { extname } from "../fs/path/mod.ts";
 import { copyN } from "../io/ioutil.ts";
 import { MultiReader } from "../io/readers.ts";
+import { FormFile } from "../multipart/formfile.ts";
+import { extname } from "../path/mod.ts";
 import { tempFile } from "../io/util.ts";
 import { BufReader, BufWriter, UnexpectedEOFError } from "../io/bufio.ts";
 import { encoder } from "../strings/mod.ts";
