+++ path/globrex.ts
