+++ path/interface.ts
