+++ path/parse_format_test.ts
@@ -3,8 +3,8 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 // TODO(kt3k): fix any types in this file
 
-import { test } from "../../testing/mod.ts";
-import { assertEquals } from "../../testing/asserts.ts";
+import { test } from "../testing/mod.ts";
+import { assertEquals } from "../testing/asserts.ts";
 import * as path from "./mod.ts";
 
 const winPaths = [
