+++ path/posix.ts
