+++ path/win32.ts
