+++ path/zero_length_strings_test.ts
@@ -2,8 +2,8 @@
 // Ported from https://github.com/browserify/path-browserify/
 
 const { cwd } = Deno;
-import { test } from "../../testing/mod.ts";
-import { assertEquals } from "../../testing/asserts.ts";
+import { test } from "../testing/mod.ts";
+import { assertEquals } from "../testing/asserts.ts";
 import * as path from "./mod.ts";
 
 const pwd = cwd();
