+++ testing/runner_test.ts
@@ -1,8 +1,8 @@
 // Copyright 2018-2019 the Deno authors. All rights reserved. MIT license.
+import { assertEquals } from "../testing/asserts.ts";
+import { isWindows } from "../path/mod.ts";
 import { test } from "./mod.ts";
 import { findTestModules } from "./runner.ts";
-import { isWindows } from "../fs/path/mod.ts";
-import { assertEquals } from "../testing/asserts.ts";
 const { cwd } = Deno;
 
 function urlToFilePath(url: URL): string {
